requirements.txt
beautifulsoup4==4.9.3
celery==5.0.0
Django==2.2.16
django-debug-toolbar==3.1.1
django-extensions==3.0.8
Faker==4.1.2
flake8==3.8.3
flake8-builtins==1.5.3
flake8-import-order==0.18.1
flake8-print==3.1.4
gunicorn==20.0.4
iniconfig==1.1.1
ipython==7.18.1
ipython-genutils==0.2.0
lxml==4.5.2
pytest==6.1.2
python-dateutil==2.8.1
python-memcached==1.59
psycopg2-binary
pytest==6.1.2
pytest-django==4.1.0
pytest-cov
pytest-mock
requests==2.24.0
uWSGI==2.0.19.1
XlsxWriter==1.3.7
